refactor(server): update makeAttendance folder service
